Evaluate Outcome Likelihood

Evaluate Outcome Likelihood

Usage

outcomeLik(
  OUTCOME,
  YEAR_FIRST,
  YEAR_LAST,
  THETA_CR,
  COVARIATES,
  NYB,
  NYA,
  YEAR_LAST_EXAM = 100L,
  LOGFLAG = FALSE
)

Arguments

OUTCOME

1 for dropout, 2 for transfer, 3 for graduation. 0 if no outcome is observed.

YEAR_FIRST

First year to evaluate.

YEAR_LAST

Last year to evaluate.

THETA_CR

Portion of the parameter vector related to the competing risk model

COVARIATES

The first 2 values refers to ability and speed respectively. Remaining values are external predictors.

NYB

Total number of years in the non-graduatable regime. Needed for determining how many time-related intercepts.

NYA

Total number of years in the graduatable regime. Needed for determining how many time-related intercepts.

YEAR_LAST_EXAM

Year at which the all exams are completed for the first time.

LOGFLAG

Set TRUE to return log value.